Adds price range criteria and code to display only 1 restaurant at a time.
After including an async waterfall in my restaurant route in Express that retrieves restaurants that meet user's criteria, parses HTML of the Yelp restaurant URLs to retrieve the price range for the restaurant, and appends that to the response, I included the front-end functionality that will enable users to select a price range.

Additionally, I removed the ng-repeat in my find-restaurant view that iterates through a list of restaurants - instead, I created a nextRestaurant() method in my findRestaurant controller that increments a number, stored in currentRestaurant. I use this number to display the next restaurant in the array when a user clicks Next Restaurant.